requirements.txt
Flask==3.0.3
psycopg2==2.9.9
python-dotenv==1.0.1
supabase==2.7.4
